requirements.txt
opencv_python==4.10.0.84
PyYAML==6.0.2
torch==2.5.1
tqdm==4.67.1
ultralytics==8.3.53
